desktop-viewer/requirements.txt
PyQt5==5.14.1
python-socketio[client]
opencv-python==4.1.2.30
opencv-contrib-python==4.1.2.30
